shelter: renderer draw quad function
Remove vertex, index buffers and shader creation in main function.
