+++ shelter/shelter/renderer.hpp
@@ -30,6 +30,7 @@ public:
 
     auto begin(camera_ref_t const& camera) -> void;
     auto submit(shader_ref_t const& shader, vertex_buffer_ref_t const& vb, index_buffer_ref_t const& ib, glm::mat4 const& model) -> void;
+    auto quad(glm::vec2 const& position, glm::vec2 const& size, glm::vec4 const& color) -> void;
     auto end() -> void;
 
     auto begin_imgui() const -> void;
@@ -37,7 +38,17 @@ public:
 
 private:
     graphics_context_ref_t m_context;
-    camera_ref_t m_camera;
+    camera_ref_t           m_camera;
+
+    struct renderer_2d{
+        shader_ref_t        shader{nullptr};
+        index_buffer_ref_t  index_buffer{nullptr};
+        vertex_buffer_ref_t vertex_buffer{nullptr};
+    };
+    renderer_2d m_2d;
+
+private:
+    auto setup_2d() -> void;
 };
 } // namespace shelter
